migrations-scripts/requirements.txt
numpy==1.26.4
pandas==2.2.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tzdata==2023.4
python-dotenv==1.0.1